Adicionado os Hash/não consegui implementar o id no Hash

+++ src/java/ModelSMBG/Persistivel.java
@@ -13,7 +13,7 @@ public class Persistivel implements Serializable{
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "id")
-    protected long id;
+    protected Integer id;
      
     @Override
     public boolean equals(Object obj) {
